requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
face-recognition==1.3.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
boto3==1.34.10
python-dotenv==1.0.0
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9

